+++ packages/extension-koni-ui/src/Popup/Home/Tokens/index.tsx
@@ -1,23 +1,22 @@
 // Copyright 2019-2022 @polkadot/extension-ui authors & contributors
 // SPDX-License-Identifier: Apache-2.0
 
-import { EmptyList, PageWrapper } from '@subwallet/extension-koni-ui/components';
-import { AccountSelectorModal } from '@subwallet/extension-koni-ui/components/Modal/AccountSelectorModal';
-import ReceiveQrModal from '@subwallet/extension-koni-ui/components/Modal/ReceiveModal/ReceiveQrModal';
-import { TokensSelectorModal } from '@subwallet/extension-koni-ui/components/Modal/ReceiveModal/TokensSelectorModal';
+import { _getOriginChainOfAsset } from '@subwallet/extension-base/services/chain-service/utils';
+import { AccountProxyType } from '@subwallet/extension-base/types';
+import { EmptyList, PageWrapper, ReceiveModal } from '@subwallet/extension-koni-ui/components';
+import BannerGenerator from '@subwallet/extension-koni-ui/components/StaticContent/BannerGenerator';
 import { TokenGroupBalanceItem } from '@subwallet/extension-koni-ui/components/TokenItem/TokenGroupBalanceItem';
 import { DEFAULT_SWAP_PARAMS, DEFAULT_TRANSFER_PARAMS, SWAP_TRANSACTION, TRANSFER_TRANSACTION } from '@subwallet/extension-koni-ui/constants';
 import { DataContext } from '@subwallet/extension-koni-ui/contexts/DataContext';
 import { HomeContext } from '@subwallet/extension-koni-ui/contexts/screen/HomeContext';
-import { useSetCurrentPage } from '@subwallet/extension-koni-ui/hooks';
+import { useCoreReceiveModalHelper, useGetBannerByScreen, useGetChainSlugsByAccount, useSetCurrentPage } from '@subwallet/extension-koni-ui/hooks';
 import useNotification from '@subwallet/extension-koni-ui/hooks/common/useNotification';
 import useTranslation from '@subwallet/extension-koni-ui/hooks/common/useTranslation';
-import useReceiveQR from '@subwallet/extension-koni-ui/hooks/screen/home/useReceiveQR';
 import { UpperBlock } from '@subwallet/extension-koni-ui/Popup/Home/Tokens/UpperBlock';
 import { RootState } from '@subwallet/extension-koni-ui/stores';
 import { ThemeProps, TransferParams } from '@subwallet/extension-koni-ui/types';
 import { TokenBalanceItemType } from '@subwallet/extension-koni-ui/types/balance';
-import { isAccountAll, sortTokenByValue } from '@subwallet/extension-koni-ui/utils';
+import { getTransactionFromAccountProxyValue, isAccountAll, sortTokenByValue } from '@subwallet/extension-koni-ui/utils';
 import { Button, Icon, SwAlert } from '@subwallet/react-ui';
 import classNames from 'classnames';
 import { Coins, FadersHorizontal } from 'phosphor-react';
@@ -36,27 +35,41 @@ const Component = (): React.ReactElement => {
   const navigate = useNavigate();
   const containerRef = useRef<HTMLDivElement>(null);
   const topBlockRef = useRef<HTMLDivElement>(null);
-  const accounts = useSelector((state: RootState) => state.accountState.accounts);
+  const accountProxies = useSelector((state: RootState) => state.accountState.accountProxies);
+  const currentAccountProxy = useSelector((state: RootState) => state.accountState.currentAccountProxy);
   const { accountBalance: { tokenGroupBalanceMap,
     totalBalanceInfo }, tokenGroupStructure: { sortedTokenGroups } } = useContext(HomeContext);
-  const currentAccount = useSelector((state: RootState) => state.accountState.currentAccount);
   const notify = useNotification();
-  const { accountSelectorItems,
-    onOpenReceive,
-    openSelectAccount,
-    openSelectToken,
-    selectedAccount,
-    selectedNetwork,
-    tokenSelectorItems } = useReceiveQR();
+  const { onOpenReceive, receiveModalProps } = useCoreReceiveModalHelper();
 
   const isZkModeSyncing = useSelector((state: RootState) => state.mantaPay.isSyncing);
   const zkModeSyncProgress = useSelector((state: RootState) => state.mantaPay.progress);
   const [, setStorage] = useLocalStorage<TransferParams>(TRANSFER_TRANSACTION, DEFAULT_TRANSFER_PARAMS);
   const [, setSwapStorage] = useLocalStorage(SWAP_TRANSACTION, DEFAULT_SWAP_PARAMS);
+  const { banners, dismissBanner, onClickBanner } = useGetBannerByScreen('token');
+  const allowedChains = useGetChainSlugsByAccount();
+  const buyTokenInfos = useSelector((state: RootState) => state.buyService.tokens);
+  const swapPairs = useSelector((state: RootState) => state.swap.swapPairs);
 
-  const transactionFromValue = useMemo(() => {
-    return currentAccount?.address ? isAccountAll(currentAccount.address) ? '' : currentAccount.address : '';
-  }, [currentAccount?.address]);
+  const fromAndToTokenMap = useMemo<Record<string, string[]>>(() => {
+    const result: Record<string, string[]> = {};
+
+    swapPairs.forEach((pair) => {
+      if (!result[pair.from]) {
+        result[pair.from] = [pair.to];
+      } else {
+        result[pair.from].push(pair.to);
+      }
+    });
+
+    return result;
+  }, [swapPairs]);
+
+  const isEnableSwapButton = useMemo(() => {
+    return Object.keys(fromAndToTokenMap).some((tokenSlug) => {
+      return allowedChains.includes(_getOriginChainOfAsset(tokenSlug));
+    });
+  }, [allowedChains, fromAndToTokenMap]);
 
   const handleScroll = useCallback((event: React.UIEvent<HTMLElement>) => {
     const topPosition = event.currentTarget.scrollTop;
@@ -132,6 +145,10 @@ const Component = (): React.ReactElement => {
 
   const isTotalBalanceDecrease = totalBalanceInfo.change.status === 'decrease';
 
+  const isSupportBuyTokens = useMemo(() => {
+    return Object.values(buyTokenInfos).some((item) => allowedChains.includes(item.network));
+  }, [allowedChains, buyTokenInfos]);
+
   const onClickItem = useCallback((item: TokenBalanceItemType) => {
     return () => {
       navigate(`/home/tokens/detail/${item.slug}`);
@@ -143,7 +160,11 @@ const Component = (): React.ReactElement => {
   }, [navigate]);
 
   const onOpenSendFund = useCallback(() => {
-    if (currentAccount && currentAccount.isReadOnly) {
+    if (!currentAccountProxy) {
+      return;
+    }
+
+    if (currentAccountProxy.accountType === AccountProxyType.READ_ONLY) {
       notify({
         message: t('The account you are using is watch-only, you cannot send assets with it'),
         type: 'info',
@@ -153,15 +174,13 @@ const Component = (): React.ReactElement => {
       return;
     }
 
-    const address = currentAccount ? isAccountAll(currentAccount.address) ? '' : currentAccount.address : '';
-
     setStorage({
       ...DEFAULT_TRANSFER_PARAMS,
-      from: address
+      fromAccountProxy: getTransactionFromAccountProxyValue(currentAccountProxy)
     });
     navigate('/transaction/send-fund');
   },
-  [currentAccount, navigate, notify, t, setStorage]
+  [currentAccountProxy, setStorage, navigate, notify, t]
   );
 
   const onOpenBuyTokens = useCallback(() => {
@@ -171,7 +190,11 @@ const Component = (): React.ReactElement => {
   );
 
   const onOpenSwap = useCallback(() => {
-    if (currentAccount && currentAccount.isReadOnly) {
+    if (!currentAccountProxy) {
+      return;
+    }
+
+    if (currentAccountProxy.accountType === AccountProxyType.READ_ONLY) {
       notify({
         message: t('The account you are using is watch-only, you cannot send assets with it'),
         type: 'info',
@@ -181,11 +204,11 @@ const Component = (): React.ReactElement => {
       return;
     }
 
-    const filteredAccounts = accounts.filter((account) => !isAccountAll(account.address));
+    const filteredAccounts = accountProxies.filter((ap) => !isAccountAll(ap.id));
 
-    const isAllLedger = (filteredAccounts.length > 0 && filteredAccounts.every((account) => account.isHardware)) || (currentAccount && !isAccountAll(currentAccount.address) && (currentAccount.isHardware));
+    const isAllLedger = currentAccountProxy.accountType === AccountProxyType.LEDGER || (filteredAccounts.length > 0 && filteredAccounts.every((ap) => ap.accountType === AccountProxyType.LEDGER));
 
-    if ((currentAccount && currentAccount.isHardware) || (isAllLedger)) {
+    if (isAllLedger) {
       notify({
         message: 'The account you are using is Ledger account, you cannot use this feature with it',
         type: 'error',
@@ -197,10 +220,10 @@ const Component = (): React.ReactElement => {
 
     setSwapStorage({
       ...DEFAULT_SWAP_PARAMS,
-      from: transactionFromValue
+      fromAccountProxy: getTransactionFromAccountProxyValue(currentAccountProxy)
     });
     navigate('/transaction/swap');
-  }, [accounts, currentAccount, navigate, notify, setSwapStorage, t, transactionFromValue]);
+  }, [accountProxies, currentAccountProxy, navigate, notify, setSwapStorage, t]);
 
   const tokenGroupBalanceItems = useMemo<TokenBalanceItemType[]>(() => {
     const result: TokenBalanceItemType[] = [];
@@ -238,6 +261,8 @@ const Component = (): React.ReactElement => {
         <UpperBlock
           isPriceDecrease={isTotalBalanceDecrease}
           isShrink={isShrink}
+          isSupportBuyTokens={isSupportBuyTokens}
+          isSupportSwap={isEnableSwapButton}
           onOpenBuyTokens={onOpenBuyTokens}
           onOpenReceive={onOpenReceive}
           onOpenSendFund={onOpenSendFund}
@@ -260,7 +285,15 @@ const Component = (): React.ReactElement => {
             />
           )
         }
-
+        {!!banners.length && (
+          <div className={'token-banner-wrapper'}>
+            <BannerGenerator
+              banners={banners}
+              dismissBanner={dismissBanner}
+              onClickBanner={onClickBanner}
+            />
+          </div>
+        )}
         {
           tokenGroupBalanceItems.map((item) => {
             return (
@@ -294,20 +327,8 @@ const Component = (): React.ReactElement => {
         </div>
       </div>
 
-      <AccountSelectorModal
-        items={accountSelectorItems}
-        onSelectItem={openSelectAccount}
-      />
-
-      <TokensSelectorModal
-        address={selectedAccount}
-        items={tokenSelectorItems}
-        onSelectItem={openSelectToken}
-      />
-
-      <ReceiveQrModal
-        address={selectedAccount}
-        selectedNetwork={selectedNetwork}
+      <ReceiveModal
+        {...receiveModalProps}
       />
     </div>
   );
@@ -320,7 +341,7 @@ const WrapperComponent = ({ className = '' }: ThemeProps): React.ReactElement<Pr
     <PageWrapper
       className={`tokens ${className}`}
       hideLoading={true}
-      resolve={dataContext.awaitStores(['price', 'chainStore', 'assetRegistry', 'balance', 'mantaPay'])}
+      resolve={dataContext.awaitStores(['price', 'chainStore', 'assetRegistry', 'balance', 'mantaPay', 'swap'])}
     >
       <Component />
     </PageWrapper>
@@ -402,6 +423,10 @@ const Tokens = styled(WrapperComponent)<ThemeProps>(({ theme: { extendToken, tok
 
     '.zk-mode-alert-area': {
       marginBottom: token.marginXS
+    },
+
+    '.token-banner-wrapper': {
+      marginBottom: token.sizeXS
     }
   });
 });