+++ packages/extension-koni-ui/src/Popup/Settings/Security/ManageWebsiteAccess/index.tsx
@@ -1,13 +1,15 @@
 // Copyright 2019-2022 @polkadot/extension-ui authors & contributors
 // SPDX-License-Identifier: Apache-2.0
 
-import { AuthUrlInfo } from '@subwallet/extension-base/background/handlers/State';
+import { AuthUrlInfo } from '@subwallet/extension-base/services/request-service/types';
+import { AccountProxy } from '@subwallet/extension-base/types';
 import { ActionItemType, ActionModal, EmptyList, FilterModal, PageWrapper, WebsiteAccessItem } from '@subwallet/extension-koni-ui/components';
 import { useDefaultNavigate, useFilterModal } from '@subwallet/extension-koni-ui/hooks';
 import { changeAuthorizationAll, forgetAllSite } from '@subwallet/extension-koni-ui/messaging';
 import { RootState } from '@subwallet/extension-koni-ui/stores';
 import { updateAuthUrls } from '@subwallet/extension-koni-ui/stores/utils';
 import { ManageWebsiteAccessDetailParam, Theme, ThemeProps } from '@subwallet/extension-koni-ui/types';
+import { isSubstrateAddress, isTonAddress } from '@subwallet/keyring';
 import { Icon, ModalContext, SwList, SwSubHeader } from '@subwallet/react-ui';
 import { FadersHorizontal, GearSix, GlobeHemisphereWest, Plugs, PlugsConnected, X } from 'phosphor-react';
 import React, { useCallback, useContext, useMemo } from 'react';
@@ -24,18 +26,30 @@ function getWebsiteItems (authUrlMap: Record<string, AuthUrlInfo>): AuthUrlInfo[
   return Object.values(authUrlMap);
 }
 
-function getAccountCount (item: AuthUrlInfo): number {
-  const authType = item.accountAuthType;
+function getAccountCount (item: AuthUrlInfo, accountProxies: AccountProxy[]): number {
+  const authType = item.accountAuthTypes;
 
-  if (authType === 'evm') {
-    return item.isAllowedMap ? Object.entries(item.isAllowedMap).filter(([address, rs]) => rs && isEthereumAddress(address)).length : 0;
+  if (!authType) {
+    return 0;
   }
 
-  if (authType === 'substrate') {
-    return item.isAllowedMap ? Object.entries(item.isAllowedMap).filter(([address, rs]) => rs && !isEthereumAddress(address)).length : 0;
-  }
+  return accountProxies.filter((ap) => {
+    return ap.accounts.some((account) => {
+      if (isEthereumAddress(account.address)) {
+        return authType.includes('evm') && item.isAllowedMap[account.address];
+      }
 
-  return Object.values(item.isAllowedMap).filter((i) => i).length;
+      if (isSubstrateAddress(account.address)) {
+        return authType.includes('substrate') && item.isAllowedMap[account.address];
+      }
+
+      if (isTonAddress(account.address)) {
+        return authType.includes('ton') && item.isAllowedMap[account.address];
+      }
+
+      return false;
+    });
+  }).length;
 }
 
 const ACTION_MODAL_ID = 'actionModalId';
@@ -50,6 +64,7 @@ enum FilterValue {
 
 function Component ({ className = '' }: Props): React.ReactElement<Props> {
   const authUrlMap = useSelector((state: RootState) => state.settings.authUrls);
+  const accountProxies = useSelector((state: RootState) => state.accountState.accountProxies);
   const { activeModal, inactiveModal } = useContext(ModalContext);
   const { t } = useTranslation();
   const navigate = useNavigate();
@@ -64,11 +79,11 @@ function Component ({ className = '' }: Props): React.ReactElement<Props> {
 
       for (const filter of selectedFilters) {
         if (filter === FilterValue.SUBSTRATE) {
-          if (item.accountAuthType === 'substrate' || item.accountAuthType === 'both') {
+          if (item.accountAuthTypes?.includes('substrate')) {
             return true;
           }
         } else if (filter === FilterValue.ETHEREUM) {
-          if (item.accountAuthType === 'evm' || item.accountAuthType === 'both') {
+          if (item.accountAuthTypes?.includes('evm')) {
             return true;
           }
         } else if (filter === FilterValue.BLOCKED) {
@@ -151,7 +166,7 @@ function Component ({ className = '' }: Props): React.ReactElement<Props> {
       navigate('/settings/dapp-access-edit', { state: {
         siteName: item.origin,
         origin: item.id,
-        accountAuthType: item.accountAuthType || ''
+        accountAuthTypes: item.accountAuthTypes || ''
       } as ManageWebsiteAccessDetailParam });
     };
   }, [navigate]);
@@ -160,7 +175,7 @@ function Component ({ className = '' }: Props): React.ReactElement<Props> {
     (item: AuthUrlInfo) => {
       return (
         <WebsiteAccessItem
-          accountCount={getAccountCount(item)}
+          accountCount={getAccountCount(item, accountProxies)}
           className={'__item'}
           domain={item.id}
           key={item.id}
@@ -169,7 +184,7 @@ function Component ({ className = '' }: Props): React.ReactElement<Props> {
         />
       );
     },
-    [onClickItem]
+    [accountProxies, onClickItem]
   );
 
   const renderEmptyList = useCallback(() => {