+++ packages/extension-koni-ui/src/types/transaction.ts
@@ -4,6 +4,7 @@
 import { StakingType } from '@subwallet/extension-base/background/KoniTypes';
 
 export interface TransactionFormBaseProps {
+  fromAccountProxy: string;
   from: string,
   chain: string
   asset: string