+++ packages/extension-web-ui/src/Popup/Transaction/variants/CancelUnstake.tsx
@@ -3,8 +3,7 @@
 
 import { _ChainInfo } from '@subwallet/chain-list/types';
 import { ExtrinsicType } from '@subwallet/extension-base/background/KoniTypes';
-import { AccountJson } from '@subwallet/extension-base/background/types';
-import { YieldPoolType, YieldPositionInfo } from '@subwallet/extension-base/types';
+import { AccountJson, YieldPoolType, YieldPositionInfo } from '@subwallet/extension-base/types';
 import { isSameAddress } from '@subwallet/extension-base/utils';
 import { AccountSelector, CancelUnstakeSelector, HiddenInput } from '@subwallet/extension-web-ui/components';
 import { ScreenContext } from '@subwallet/extension-web-ui/contexts/ScreenContext';